Move common case to tfjs-core tests
